RAG/requirements.txt
# IPN RAG Chatbot Dependencies

# Flask API
flask==3.0.3
flask-cors==4.0.1

# Vector Store & Embeddings
faiss-cpu==1.13.2
sentence-transformers>=3.0.0
numpy>=1.26.0

# LLM Integration
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-groq==0.1.6

# Document Processing
# Note: Using built-in markdown/text loaders from langchain-community

# Environment & Utilities
python-dotenv==1.0.1
pydantic==2.8.2

# Optional: For production deployment
gunicorn==22.0.0
